requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy==2.0.*
pydantic==2.10.*
websockets==14.*
jinja2==3.*
python-multipart==0.*
bcrypt>=4.0
python-jose[cryptography]==3.5.*
